requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
backports.zstd==1.0.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
build==1.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.4
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
fonttools==4.60.1
future==1.0.0
greenlet==3.2.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.3.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
mistune==3.1.4
narwhals==2.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
pillow==11.3.0
pip-tools==7.5.0
pipreqs
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pyodbc==5.2.0
pyparsing==3.2.5
pyproject_hooks==1.2.0
pyspnego==0.12.0
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pywinrm==0.5.0
pyzk==0.9
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
requests_ntlm==1.3.0
retrying==1.4.2
rpds-py==0.27.1
setuptools==75.9.1
six==1.17.0
soupsieve==2.8
speedtest-cli==2.1.3
sspilib==0.4.0
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
xmltodict==0.15.1
yarg==0.1.9
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1